Introduce main storages "cache" and "store"
I first thought storages will simply be registered as a hash, and that
users will always specify which storage they want. But this lowers the
convenience, especially when Attachable module is introduced and we have
to specify both cache and store storages explicitly. Instead, we make
"cache" and "store" storages the main ones, since this is the most
common usage of uploading libraries. Then we will have Attachable (and
potentially some other plugins) default to these storages.