Rewind file after uploading
After uploading we're also potentially extracting metadata, so the file
should be in the same state as before uploading, because the user
shouldn't have to know whether they're extracting metadata before or
after uploading. Later we will probably move extracting metadata before
uploading, so this ensures we won't break anything with that change.