Storages return URL after uploading
I think this is a good convention, because it can happen that we have a
storage which requires an HTTP request to generate the URL (e.g.
Dropbox), and these service's uploading endpoint may already return the
URL as the response, so maybe in a future situation this would save an
HTTP request.