Add validation to core
The reason why we're adding it to core is that the upcoming Attachable
can already count on it, as well as the ActiveRecord and Sequel
extensions. It would be difficult from a validation plugin to let them
know that validations exist. This way we also have a convenient
`validate` instance method which we can use without loading helpers.