Add <attachment>_url
This is useful for APIs, so that we can reference the URL by a single
method name. It's also useful as convenient method that return nil when
there is no attachment (as opposed to having to do `user.avatar.url if
user.avatar`), and also gets in handy when combined with the versions
plugin. Finally, the upcoming default URL plugin can now use this
method.