Split processing and versions again
Often people will want to process files without generating versions, and
in that case they shouldn't have to load all the code for managing
versions. We make the versions plugin a "private" plugin, because it
doesn't really make sense to use it on its own without processing, so it
would be confusing to have it as a feature.