Make soft_delete work only on #destroy
When people say "soft delete", they mean a scenario where instead of
deleting a record, the record is set a "deleted_at" column (so that it
can be potentially be undeleted later), and in that case they want to
preserve the attachment. So the soft_delete plugin should do exactly
that, simply not delete the file when record is scheduled for
destruction.

There is a possibility later that we will add to this plugin the
funcionality to not delete replaced or nullified files, but only if this
feature gets requested.