Catch InvalidFile errors in endpoint plugin
We don't want that attackers are able to cause errors to your website
simply by passing something that isn't a file. There is no reason for
this error to propagate, if developers are doing something wrong they
will see that they got a 400 error response instead of 500, so they
should check the response body.

And btw, "Bad Request" is a 400, not a 500 :P