Implement the concept of moving files
I believe this is a unique feature compared to other uploading
solutions. When you look at the FileSystem, moving files from one place
to the other doesn't effectively take any time (internally the operating
system is just moving the pointer). This doesn't make much difference
with small files, but it can make with larger files (e.g. videos). This
means that if the user uses FileSystem for both storages, there isn't
any copying whatsoever.

This concept can be expanded to other storages as well. For example,
Dropbox has the option of moving files, so if the user has both cache
and store Dropbox, then moving the file from cache and store requires
only one HTTP request (no uploading needs to be done).

If the storage doesn't support moving in general, or a given IO cannot
be moved (e.g. cache is FileSystem and store is S3), then we imitate
moving by uploading the file and then deleting it afterwards. The
"delete_cached" plugin is a subset of this feature, so we're removing
it.

I believe this concept is very powerful, and that a lot of users may
find it useful for various scenarios.