Add the delete_invalid plugin
Some people might have a problem with invalid files staying cached, so
for them we offer this plugin. The reason why this is not the default is
because I believe in some scenarios this behaviour can surprise users.
For example, when using the Endpoint plugin and caching the file, it may
happen that the user wants to reference the cached file again later in
the controller, and then there is the contradiction that you have an
UploadedFile which is not actually uploaded.

For these reasons I feel safer that this behaviour is opt-in rather than
opt-out. I don't think that it should bother a lot of people that
invalid files stay cached, because the cache should be cleaned regularly
anyway, and invalid files in my experience aren't that common in
general. And if the scenario is a bit more delicate, it's better to do
JavaScript file validations.