Add the :max_size options to endpoint plugin
Since users can't control which filesizes they accept (like they can in
the attacher), it is good to provide a security check just to make sure
there is no deliberate attempts of overflowing the cache. We also delete
the file from cache, because there is a chance that Ruby Tempfiles are
deleted only when the Ruby process is terminated.