No need to allow deleting cache in background
Deleting cache only happens with the "moving" plugin, but then it's part
of "promoting". If a user decides to put promoting in a background job,
then it makes sense that deleting cache is also part of that job
(instead of that job firing another worker).

This now greatly simplifies the implementation, so now we can treat
deleting in the same level as promoting. It can now be part of
Sequel/ActiveRecord plugins.