Refactor to allow processing files in any phase
We initially thought that users will only want to process on caching or
on storing. But there are actually 4 points in time where users may want to
process files:

1. On direct upload to the endpoint
2. On assignment after regular form submission
3. When saving, but before promoting
4. When promoting

None of these points are exclusive, users may want to do processing on 1
or more of these points. For example, they want to normalize the file on
direct upload, generate some versions on saving, and then generate the
rest of the versions in a background job during promoting. That's why we
introduce the concept of "phases", to allow users to do processing in
any of them if they wish to.

By making a single #process instance method, we give the benefit of
having all processing code in one place, which also means that we can
implement logging easily.

We also added here the method `#path`, which returns non-nil only if
the storage is FileSystem. This allows users to choose to do the
processing in a way that no tempfiles are left over, enforcing our
ability to leave no garbage.