Make Shrine::Utils#download more robust
It can now catch invalid URLs, and transform them to Shrine::Error.
Furthermore, it will not swallow all RuntimeErrors now, only those which
were raised by open-uri as redirection loops.

We also give it the ability to accept :max_size. This is really great,
because open-uri allows you to pass a proc which will get called as soon
as it gets the "Content-Length" header. This means that as soon as we
realize that the file will be too big, we can terminate the download.